requirements.txt
pyarrow>=12.0
yt-dlp>=2024.1.1
psycopg2-binary>=2.9
fastapi>=0.110
uvicorn>=0.27
